[sanitizer_symbolizer] Split Fuchsia and Markup.
This PR separated parts of the symbolizer markup
implementation that are Fuchsia OS specific. This
is in preparation of enabling symbolizer markup
in other OSs.
